Socket build flag replaced with endpoint check
It will now attempt to connect to a unix socket if the given
HSDS_ENDPOINT begins with "http+unix", instead of
requiring a particular build option.

Currently, connecting to a unix socket requires
username and other auth info to be in environment variables
(as opposed to in a config file).