Fix empty request on 0-sized write
REST VOL used to send a PUT request with an empty body when a 0-element
write occurred, which would result in a 400 error from the server.

Now it skips the request if the body is size 0.

A more involved workaround may be needed if parallelism is
implemented, since coordination might rely on each worker
completing an I/O request.