Fix read callback to handle >64kb transfers
curl uses an internal buffer of size 64kb. When the VOL had a payload
larger than that to upload, the read data callback would copy the first
64kb of data each time, and never return 0 to indicate the transfer
was finished. It would continue to write to the server until
the server closed its end of the connection.

Also fixed the read large point selection test to no longer
access invalid memory. If run, it now receives a
413 Entity too large error from HSDS instead.