Enable additional H5Aopen* calls in test_rest_vol
A couple calls to H5Aopen_by_idx in test_rest_vol.c provided an
erroneous relative path, so I replaced those paths with "."

Specifically, the erroneous calls provided a datatype as a
parent object, then provided the name of that datatype as a relative
path.  Since the datatype doesn't contain itself, this would always
lead to the object not being found.

I also modified H5Aopen_by_idx to not make an unnecessary request to
the server when the relative path is ".", because then the provided
parent object is the same as the target object, and no path traversal
is needed.